Break example.c into a synchronous and an asynchronous flavor
