+++ examples/example.c
@@ -1,6 +1,6 @@
 /*********************************************************************
  *
- * example.c: example Riak C Client application
+ * example.c: Rxample Riak C Client Synchronous Application
  *
  * Copyright (c) 2007-2014 Basho Technologies, Inc.  All Rights Reserved.
  *
@@ -21,77 +21,8 @@
  *********************************************************************/
 
 #include "riak.h"
-#include "riak_async.h"
 #include "riak_command.h"
-#include "example_call_backs.h"
-#include "../adapters/riak_libevent.h"
-#include <time.h>
-
-
-//********************************************************************/
-// Logging struct(s)/functions
-//********************************************************************/
-typedef struct {
-    FILE *fp;
-} example_log_data;
-
-// log initialization callback, open files here etc.
-riak_int32_t
-example_log_init(void *ptr) {
-    example_log_data* datum = (example_log_data*)ptr;
-    datum->fp = fopen("riak.log", "w+");
-    if (datum->fp == NULL) {
-        return 1;
-    }
-    return 0;
-}
-
-// log cleanup callback, close files here etc.
-void
-example_log_cleanup(void *ptr) {
-    example_log_data* datum = (example_log_data*)ptr;
-    fclose(datum->fp);
-}
-
-// handle an incoming log message
-void
-example_log(void            *ptr,
-            riak_log_level_t level,
-            const char      *file,
-            riak_size_t      filelen,
-            const char      *func,
-            riak_size_t      funclen,
-            riak_uint32_t    line,
-            const char      *format,
-            va_list          args) {
-    example_log_data* datum = (example_log_data*)ptr;
-    time_t ltime;
-    struct tm result;
-    char stime[32];
-    char output[2048];
-
-    if (datum->fp == NULL) return;
-
-    ltime = time(NULL);
-    localtime_r(&ltime, &result);
-    strftime(stime, sizeof(stime), "%F %X %Z", &result);
-    riak_size_t wrote = snprintf(output, sizeof(output), "%s %s ", stime, riak_log_level_description(level));
-    char *pos = output + wrote;
-    wrote += vsnprintf(pos, sizeof(output)-wrote, format, args);
-    if (wrote < sizeof(output)-1) {
-        strcat(output, "\n");
-    }
-    fprintf(datum->fp, "%s", output);
-    if (level == RIAK_LOG_CRITICAL) {
-        fprintf(stderr, "%s", output);
-    }
-}
-
-
-
-//********************************************************************/
-// main application
-//********************************************************************/
+#include "example_log.h"
 
 int
 main(int   argc,
@@ -102,10 +33,6 @@ main(int   argc,
     if (operation < 0) {
         usage(stderr, argv[0]);
     }
-
-#ifdef _RIAK_DEBUG
-    event_enable_debug_mode();
-#endif
 
     // a riak_config serves as your per-thread state to interact with Riak
     riak_config *cfg;
@@ -126,32 +53,10 @@ main(int   argc,
         exit(1);
     }
 
-
-    riak_object *obj;
-    riak_bucketprops *props;
-    riak_get_options *get_options;
-    riak_delete_options *delete_options;
-    riak_2index_options *index_options;
-    riak_search_options *search_options;
     char output[10240];
     riak_print_state print_state;
     riak_print_init(&print_state, output, sizeof(output));
 
-    int result = evthread_use_pthreads();
-    if (result < 0) {
-        riak_log_critical_config(cfg, "%s", "Could not use pthreads with libevent");
-        exit(1);
-    }
-    struct event_base *base = event_base_new();
-    if (base == NULL) {
-        riak_log_critical_config(cfg, "%s", "Could not create libevent base");
-        exit(1);
-    }
-    result = evthread_make_base_notifiable(base);
-    if (result) {
-        riak_log_critical_config(cfg, "%s", "Could not initialize libevent base");
-        exit(1);
-    }
     int it;
 
     // create some sample binary values to use
@@ -169,11 +74,32 @@ main(int   argc,
         exit(1);
     }
 
+    // Supporting Options and outputs
+    riak_2index_options *index_options;
+    riak_bucketprops    *props;
+    riak_delete_options *delete_options;
+    riak_get_options    *get_options;
+    riak_object         *obj;
+    riak_put_options    *put_options;
+    riak_search_options *search_options;
+
+    // Every possible message response type
+    riak_2index_response            *index_response = NULL;
+    riak_get_bucketprops_response   *props_response = NULL;
+    riak_get_clientid_response      *getcli_response = NULL;
+    riak_get_response               *get_response = NULL;
+    riak_listbuckets_response       *bucket_response = NULL;
+    riak_listkeys_response          *key_response = NULL;
+    riak_mapreduce_response         *mapred_response = NULL;
+    riak_put_response               *put_response = NULL;
+    riak_reset_bucketprops_response *reset_response = NULL;
+    riak_search_response            *search_response = NULL;
+    riak_serverinfo_response        *serverinfo_response = NULL;
+    riak_set_clientid_response      *setcli_response = NULL;
+
     // iterate through argv
     for(it = 0; it < args.iterate; it++) {
         riak_connection  *cxn   = NULL;
-        riak_libevent    *event = NULL;
-        riak_operation   *rop   = NULL;
         // Create a connection with the default address resolver
         err = riak_connection_new(cfg, &cxn, args.host, args.portnum, NULL);
         if (err) {
@@ -182,46 +108,22 @@ main(int   argc,
         // if the application was configured using  -D_RIAK_DEBUG,
         // it will trace debug level messages
         riak_log_debug(cxn, "Loop %d", it);
-        err = riak_operation_new(cxn, &rop, NULL, NULL, NULL);
-        if (err) {
-            fprintf(stderr, "Could not allocate operation\n");
-            exit(1);
-        }
-        // TODO: Operation callbacks?  Needed at construction time?
-
-        if (args.async) {
-            err = riak_libevent_new(&event, rop, base);
-            if (err) {
-                return err;
-            }
-            riak_operation_set_error_cb(rop, example_error_cb);
-            riak_operation_set_cb_data(rop, rop);
-        }
 
         // handle possible operations from the command line
         switch (operation) {
         case RIAK_COMMAND_PING:
-            if (args.async) {
-                err = riak_async_register_ping(rop, (riak_response_callback)example_ping_cb);
-            } else {
-                err = riak_ping(cxn);
-                printf("PONG\n");
-            }
+            err = riak_ping(cxn);
+            printf("PONG\n");
             if (err) {
                 fprintf(stderr, "No Ping [%s]\n", riak_strerror(err));
                 exit(1);
             }
             break;
         case RIAK_COMMAND_GETSERVERINFO:
-            if (args.async) {
-                err = riak_async_register_serverinfo(rop, (riak_response_callback)example_serverinfo_cb);
-            } else {
-                riak_serverinfo_response *serverinfo_response = NULL;
-                err = riak_serverinfo(cxn, &serverinfo_response);
-                riak_serverinfo_response_print(&print_state, serverinfo_response);
-                printf("%s\n", output);
-                riak_serverinfo_response_free(cfg, &serverinfo_response);
-            }
+            err = riak_serverinfo(cxn, &serverinfo_response);
+            riak_serverinfo_response_print(&print_state, serverinfo_response);
+            printf("%s\n", output);
+            riak_serverinfo_response_free(cfg, &serverinfo_response);
             if (err) {
                 fprintf(stderr, "Server Info Problems [%s]\n", riak_strerror(err));
                 exit(1);
@@ -235,17 +137,12 @@ main(int   argc,
             }
             riak_get_options_set_basic_quorum(get_options, RIAK_TRUE);
             riak_get_options_set_r(get_options, 2);
-            if (args.async) {
-                err = riak_async_register_get(rop, bucket_bin, key_bin, get_options, (riak_response_callback)example_get_cb);
-            } else {
-                riak_get_response *get_response = NULL;
-                err = riak_get(cxn, bucket_bin, key_bin, get_options, &get_response);
-                if (err == ERIAK_OK) {
-                    riak_get_response_print(&print_state, get_response);
-                    printf("%s\n", output);
-                }
-                riak_get_response_free(cfg, &get_response);
+            err = riak_get(cxn, bucket_bin, key_bin, get_options, &get_response);
+            if (err == ERIAK_OK) {
+                riak_get_response_print(&print_state, get_response);
+                printf("%s\n", output);
             }
+            riak_get_response_free(cfg, &get_response);
             riak_get_options_free(cfg, &get_options);
             if (err) {
                 fprintf(stderr, "Get Problems [%s]\n", riak_strerror(err));
@@ -269,7 +166,7 @@ main(int   argc,
                 riak_free(cfg, &obj);
                 exit(1);
             }
-            riak_put_options *put_options = riak_put_options_new(cfg);
+            put_options = riak_put_options_new(cfg);
             if (put_options == NULL) {
                 riak_log_critical(cxn, "%s","Could not allocate a Riak Put Options");
                 return 1;
@@ -277,17 +174,12 @@ main(int   argc,
             riak_put_options_set_return_head(put_options, RIAK_TRUE);
             riak_put_options_set_return_body(put_options, RIAK_TRUE);
 
-            if (args.async) {
-                err = riak_async_register_put(rop, obj, put_options, (riak_response_callback)example_put_cb);
-            } else {
-                riak_put_response *put_response = NULL;
-                err = riak_put(cxn, obj, put_options, &put_response);
-                if (err == ERIAK_OK) {
-                    riak_put_response_print(&print_state, put_response);
-                    printf("%s\n", output);
-                }
-                riak_put_response_free(cfg, &put_response);
+            err = riak_put(cxn, obj, put_options, &put_response);
+            if (err == ERIAK_OK) {
+                riak_put_response_print(&print_state, put_response);
+                printf("%s\n", output);
             }
+            riak_put_response_free(cfg, &put_response);
             riak_object_free(cfg, &obj);
             riak_put_options_free(cfg, &put_options);
             if (err) {
@@ -303,11 +195,7 @@ main(int   argc,
             }
             riak_delete_options_set_w(delete_options, 1);
             riak_delete_options_set_dw(delete_options, 1);
-           if (args.async) {
-                err = riak_async_register_delete(rop, bucket_bin, key_bin, delete_options, (riak_response_callback)example_delete_cb);
-            } else {
-                err = riak_delete(cxn, bucket_bin, key_bin, delete_options);
-            }
+            err = riak_delete(cxn, bucket_bin, key_bin, delete_options);
             riak_delete_options_free(cfg, &delete_options);
             if (err) {
                 fprintf(stderr, "Delete Problems [%s]\n", riak_strerror(err));
@@ -315,95 +203,65 @@ main(int   argc,
             }
             break;
         case RIAK_COMMAND_LISTBUCKETS:
-            if (args.async) {
-                err = riak_async_register_listbuckets(rop, (riak_response_callback)example_listbucket_cb);
-            } else {
-                riak_listbuckets_response *bucket_response = NULL;
-                err = riak_listbuckets(cxn, &bucket_response);
-                if (err == ERIAK_OK) {
-                    riak_listbuckets_response_print(&print_state, bucket_response);
-                    printf("%s\n", output);
-                }
-                riak_listbuckets_response_free(cfg, &bucket_response);
+            err = riak_listbuckets(cxn, &bucket_response);
+            if (err == ERIAK_OK) {
+                riak_listbuckets_response_print(&print_state, bucket_response);
+                printf("%s\n", output);
             }
+            riak_listbuckets_response_free(cfg, &bucket_response);
             if (err) {
                 fprintf(stderr, "List Buckets Problems [%s]\n", riak_strerror(err));
                 exit(1);
             }
             break;
         case RIAK_COMMAND_LISTKEYS:
-            if (args.async) {
-                err = riak_async_register_listkeys(rop, bucket_bin, args.timeout * 1000, (riak_response_callback)example_listkey_cb);
-            } else {
-                riak_listkeys_response *key_response = NULL;
-                err = riak_listkeys(cxn, bucket_bin, args.timeout * 1000, &key_response);
-                if (err == ERIAK_OK) {
-                    riak_listkeys_response_print(&print_state, key_response);
-                    printf("%s\n", output);
-                }
-                riak_listkeys_response_free(cfg, &key_response);
+            err = riak_listkeys(cxn, bucket_bin, args.timeout * 1000, &key_response);
+            if (err == ERIAK_OK) {
+                riak_listkeys_response_print(&print_state, key_response);
+                printf("%s\n", output);
             }
+            riak_listkeys_response_free(cfg, &key_response);
             if (err) {
                 fprintf(stderr, "List Keys Problems [%s]\n", riak_strerror(err));
                 exit(1);
             }
             break;
         case RIAK_COMMAND_GETCLIENTID:
-            if (args.async) {
-                err = riak_async_register_get_clientid(rop, (riak_response_callback)example_getclientid_cb);
-            } else {
-                riak_get_clientid_response *getcli_response = NULL;
-                err = riak_get_clientid(cxn, &getcli_response);
-                if (err == ERIAK_OK) {
-                    riak_get_clientid_response_print(&print_state, getcli_response);
-                    printf("%s\n", output);
-                }
-                riak_get_clientid_response_free(cfg, &getcli_response);
+            err = riak_get_clientid(cxn, &getcli_response);
+            if (err == ERIAK_OK) {
+                riak_get_clientid_response_print(&print_state, getcli_response);
+                printf("%s\n", output);
             }
+            riak_get_clientid_response_free(cfg, &getcli_response);
             if (err) {
                 fprintf(stderr, "Get ClientID Problems [%s]\n", riak_strerror(err));
                 exit(1);
             }
             break;
         case RIAK_COMMAND_SETCLIENTID:
-            if (args.async) {
-                err = riak_async_register_set_clientid(rop, value_bin, (riak_response_callback)example_setclientid_cb);
-            } else {
-                riak_set_clientid_response *setcli_response = NULL;
-                err = riak_set_clientid(cxn, value_bin, &setcli_response);
-                riak_set_clientid_response_free(cfg, &setcli_response);
-            }
+            err = riak_set_clientid(cxn, value_bin, &setcli_response);
+            riak_set_clientid_response_free(cfg, &setcli_response);
             if (err) {
                 fprintf(stderr, "Set ClientID Problems [%s]\n", riak_strerror(err));
                 exit(1);
             }
             break;
         case RIAK_COMMAND_GETBUCKET:
-            if (args.async) {
-                err = riak_async_register_get_bucketprops(rop, bucket_bin, (riak_response_callback)example_getbucketprops_cb);
-            } else {
-                riak_get_bucketprops_response *bucket_response = NULL;
-                err = riak_get_bucketprops(cxn, bucket_bin, &bucket_response);
-                if (err == ERIAK_OK) {
-                    riak_get_bucketprops_response_print(&print_state, bucket_response);
-                    printf("%s\n", output);
-                }
-                riak_get_bucketprops_response_free(cfg, &bucket_response);
+            err = riak_get_bucketprops(cxn, bucket_bin, &props_response);
+            if (err == ERIAK_OK) {
+                riak_get_bucketprops_response_print(&print_state, props_response);
+                printf("%s\n", output);
             }
+            riak_get_bucketprops_response_free(cfg, &props_response);
             if (err) {
                 fprintf(stderr, "Get Bucket Properties Problems [%s]\n", riak_strerror(err));
                 exit(1);
             }
             break;
-            case RIAK_COMMAND_RESETBUCKET:
-            if (args.async) {
-                err = riak_async_register_reset_bucketprops(rop, bucket_bin, (riak_response_callback)example_resetbucketprops_cb);
-            } else {
-                riak_reset_bucketprops_response *bucket_response = NULL;
-                err = riak_reset_bucketprops(cxn, bucket_bin, &bucket_response);
-                if (err == ERIAK_OK) {
-                    riak_reset_bucketprops_response_free(cfg, &bucket_response);
-                }
+        case RIAK_COMMAND_RESETBUCKET:
+            err = riak_reset_bucketprops(cxn, bucket_bin, &reset_response);
+            if (err == ERIAK_OK) {
+                riak_reset_bucketprops_response_free(cfg, &reset_response);
             }
             if (err) {
                 fprintf(stderr, "Reset Bucket Properties Problems [%s]\n", riak_strerror(err));
@@ -417,32 +275,25 @@ main(int   argc,
                 return 1;
             }
             riak_bucketprops_set_last_write_wins(props, RIAK_FALSE);
-            if (args.async) {
-                err = riak_async_register_set_bucketprops(rop, bucket_bin, props, (riak_response_callback)example_setbucketprops_cb);
-            } else {
-                riak_set_bucketprops_response *bucket_response = NULL;
-                 err = riak_set_bucketprops(cxn, bucket_bin, props, &bucket_response);
-                 if (err == ERIAK_OK) {
-                     riak_set_bucketprops_response_free(cfg, &bucket_response);
-                 }
-            }
+            riak_set_bucketprops_response *bucket_response = NULL;
+             err = riak_set_bucketprops(cxn, bucket_bin, props, &bucket_response);
+             if (err == ERIAK_OK) {
+                 riak_set_bucketprops_response_free(cfg, &bucket_response);
+             }
             if (err) {
                 fprintf(stderr, "Set Bucket Properties Problems [%s]\n", riak_strerror(err));
                 exit(1);
             }
             break;
         case RIAK_COMMAND_MAPRED:
-            if (args.async) {
-                err = riak_async_register_mapreduce(rop, content_type, value_bin, RIAK_FALSE, (riak_response_callback)example_mapreduce_cb);
-            } else {
-                riak_mapreduce_response *mapred_response = NULL;
-                err = riak_mapreduce(cxn, content_type, value_bin, &mapred_response);
-                if (err == ERIAK_OK) {
-                    riak_mapreduce_response_print(&print_state, mapred_response);
-                    printf("%s\n", output);
-                }
-                riak_mapreduce_response_free(cfg, &mapred_response);
-            }
+
+             err = riak_mapreduce(cxn, content_type, value_bin, &mapred_response);
+             if (err == ERIAK_OK) {
+                 riak_mapreduce_response_print(&print_state, mapred_response);
+                 printf("%s\n", output);
+
+                 riak_mapreduce_response_free(cfg, &mapred_response);
+             }
             if (err) {
                 fprintf(stderr, "Map/Reduce Problems [%s]\n", riak_strerror(err));
                 exit(1);
@@ -457,17 +308,12 @@ main(int   argc,
             riak_2index_options_set_stream(index_options, RIAK_TRUE);
             riak_2index_options_set_timeout(index_options, 10000);
             riak_2index_options_set_key(cfg, index_options, value_bin);
-            if (args.async) {
-                err = riak_async_register_2index(rop, bucket_bin, index_bin, index_options, (riak_response_callback)example_2index_cb);
-            } else {
-                riak_2index_response *index_response = NULL;
-                err = riak_2index(cxn, bucket_bin, index_bin, index_options, &index_response);
-                if (err == ERIAK_OK) {
-                    riak_2index_response_print(&print_state, index_response);
-                    printf("%s\n", output);
-                }
-                riak_2index_response_free(cfg, &index_response);
+            err = riak_2index(cxn, bucket_bin, index_bin, index_options, &index_response);
+            if (err == ERIAK_OK) {
+                riak_2index_response_print(&print_state, index_response);
+                printf("%s\n", output);
             }
+            riak_2index_response_free(cfg, &index_response);
             riak_2index_options_free(cfg, &index_options);
             if (err) {
                 fprintf(stderr, "Secondary Index Problems [%s]\n", riak_strerror(err));
@@ -482,17 +328,12 @@ main(int   argc,
             }
             riak_search_options_set_start(search_options, 0);
             riak_search_options_set_rows(search_options, 50);
-            if (args.async) {
-                err = riak_async_register_search(rop, bucket_bin, value_bin, search_options, (riak_response_callback)example_search_cb);
-            } else {
-                riak_search_response *search_response = NULL;
-                err = riak_search(cxn, bucket_bin, value_bin, search_options, &search_response);
-                if (err == ERIAK_OK) {
-                    riak_search_response_print(&print_state, search_response);
-                    printf("%s\n", output);
-                }
-                riak_search_response_free(cfg, &search_response);
+            err = riak_search(cxn, bucket_bin, value_bin, search_options, &search_response);
+            if (err == ERIAK_OK) {
+                riak_search_response_print(&print_state, search_response);
+                printf("%s\n", output);
             }
+            riak_search_response_free(cfg, &search_response);
             riak_search_options_free(cfg, &search_options);
             if (err) {
                 fprintf(stderr, "Search Problems [%s]\n", riak_strerror(err));
@@ -501,24 +342,11 @@ main(int   argc,
             break;
         default:
             usage(stderr, argv[0]);
-        }
-
-        if (args.async) {
-            err = riak_libevent_send(rop, event);
-            if (err) {
-                riak_log_critical(cxn, "%s", "Could not send request");
-                exit(1);
-            }
-        }
-    }
-
-    if (args.async) {
-        // Terminates only on error or timeout
-        event_base_dispatch(base);
+            exit(1);
+       }
     }
 
     // cleanup
-    event_base_free(base);
     riak_binary_free(cfg, &bucket_bin);
     riak_binary_free(cfg, &key_bin);
     riak_binary_free(cfg, &value_bin);
